requirements.txt
pandas==2.1.4
numpy==1.24.3
apache-airflow==2.7.3
apache-kafka==3.6.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
boto3==1.34.0
azure-storage-blob==12.19.0
google-cloud-storage==2.10.0
streamlit==1.28.1
plotly==5.17.0
dash==2.14.2
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
schedule==1.2.0
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
redis==5.0.1
celery==5.3.4
flower==2.0.1
prometheus-client==0.19.0
structlog==23.2.0
pytest==7.4.3
black==23.11.0
flake8==6.1.0
